+++ packages/ui/client/components.d.ts
@@ -7,7 +7,7 @@ declare module 'vue' {
     CodeMirror: typeof import('./components/CodeMirror.vue')['default']
     ConnectionOverlay: typeof import('./components/ConnectionOverlay.vue')['default']
     Dashboard: typeof import('./components/Dashboard.vue')['default']
-    DashboardEntry: typeof import('./components/DashboardEntry.vue')['default']
+    DashboardEntry: typeof import('./components/dashboard/DashboardEntry.vue')['default']
     DetailsPanel: typeof import('./components/DetailsPanel.vue')['default']
     FileDetails: typeof import('./components/FileDetails.vue')['default']
     IconButton: typeof import('./components/IconButton.vue')['default']
@@ -20,8 +20,9 @@ declare module 'vue' {
     TaskItem: typeof import('./components/TaskItem.vue')['default']
     TasksList: typeof import('./components/TasksList.vue')['default']
     TaskTree: typeof import('./components/TaskTree.vue')['default']
-    TestEntry: typeof import('./components/TestEntry.vue')['default']
-    TestFilesEntry: typeof import('./components/TestFilesEntry.vue')['default']
+    TestFilesEntry: typeof import('./components/dashboard/TestFilesEntry.vue')['default']
+    TestsEntry: typeof import('./components/dashboard/TestsEntry.vue')['default']
+    TestsFilesContainer: typeof import('./components/dashboard/TestsFilesContainer.vue')['default']
     ViewEditor: typeof import('./components/views/ViewEditor.vue')['default']
     ViewModuleGraph: typeof import('./components/views/ViewModuleGraph.vue')['default']
     ViewReport: typeof import('./components/views/ViewReport.vue')['default']
